removes iconcomponent
