+++ Robust.Server/GameObjects/ServerComponentFactory.cs
@@ -29,7 +29,6 @@ namespace Robust.Server.GameObjects
             Register<MapGridComponent>();
             RegisterReference<MapGridComponent, IMapGridComponent>();
 
-            RegisterIgnore("Icon");
             RegisterIgnore("Eye");
 
             Register<BasicActorComponent>();
